requirements.txt
flask==3.0.3
requests==2.31.0
beautifulsoup4==4.12.3
Pillow==10.3.0
pillow-avif-plugin==1.4.6
lxml==5.2.2
python-dotenv==1.0.1
playwright==1.58.0
APScheduler==3.10.4
